requirements.txt
Cython
Django==3.2
django-q==1.3.9

# Spacy & Models
spacy==3.7.4
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889

# Google AI & API (Fixing Dependency Conflict)
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.127.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.62.0

# LangChain & Related
langchain
langchain_community
langsmith
langchain_google_genai
chromadb==0.5.23

# NLP & Data Processing
nltk==3.8.1
pandas==2.2.1
PyMuPDF==1.23.6
pypdf==4.2.0
rank-bm25==0.2.2
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.14.1
sentence-transformers==2.6.1
setuptools==69.1.0
transformers==4.38.2
textblob==0.18.0
beautifulsoup4==4.12.3
python-decouple==3.8
wheel==0.43.0
